Refactor 'commit' command to new pattern
This update refactors the 'commit' command to follow a new pattern. It introduces a new `handler.ts` file which contains the main logic for the commit command. The `index.ts` file has been added to export the command, description, builder, handler and options. The `options.ts` file contains the command line options via yargs. The old `commit.ts` file has been removed as it's no longer needed.
