+++ src/commands/commit/index.ts
@@ -1,0 +1,10 @@
+import { handler } from './handler'
+import { builder, options } from './options'
+
+export default {
+  command: 'commit',
+  desc: 'Generate commit message',
+  builder,
+  handler,
+  options,
+}
